# swagger/swagger.yaml
basePath: /
definitions:
...
  logtrace.Organization:
    properties:
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      is_active:
        type: boolean
      is_subscription_active:
        type: boolean
      name:
        type: string
      plan_name:
        type: string
      updated_at:
        type: string
    type: object
...
  server.APIStatus:
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  server.OrganizationResponse:
    properties:
      message:
        type: string
      organization:
        $ref: '#/definitions/logtrace.Organization'
    required:
    - message
    type: object
...
  server.createInvitationRequest:
    properties:
      email:
        type: string
      fullname:
        type: string
      role:
        type: string
    type: object
  server.createOrgRequest:
    properties:
      image_url:
        type: string
      name:
        type: string
    type: object
...
  server.updateOrgRequest:
    properties:
      image_url:
        type: string
      is_active:
        type: boolean
      is_subscription_active:
        type: boolean
      name:
        type: string
      plan_name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Logtrace API documentation
  title: Logtrace API
  version: 0.1.0
paths:
...
  /v1/invitations:
    post:
      consumes:
      - application/json
      description: Create a new invitation
      parameters:
      - description: Invitation creation request
        in: body
        name: invitation
        required: true
        schema:
          $ref: '#/definitions/server.createInvitationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Invitation created successfully
          schema:
            $ref: '#/definitions/server.APIStatus'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/server.APIStatus'
        "500":
          description: Could not create invitation at this time. an error occurred
          schema:
            $ref: '#/definitions/server.APIStatus'
      tags:
      - Invitations
  /v1/invitations/{token}/accept:
    post:
      consumes:
      - application/json
      description: Accept an invitation by token
      parameters:
      - description: Invitation token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invitation accepted successfully
          schema:
            $ref: '#/definitions/server.APIStatus'
        "400":
          description: Invalid invitation token
          schema:
            $ref: '#/definitions/server.APIStatus'
        "404":
          description: Invitation not found
          schema:
            $ref: '#/definitions/server.APIStatus'
        "500":
          description: Failed to accept invitation
          schema:
            $ref: '#/definitions/server.APIStatus'
      tags:
      - Invitations
  /v1/organizations:
    post:
      consumes:
      - application/json
      description: Create a new organization
      parameters:
      - description: Organization creation request
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/server.createOrgRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Organization created successfully
          schema:
            $ref: '#/definitions/server.APIStatus'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/server.APIStatus'
        "500":
          description: Could not create organization at this time. an error occurred
          schema:
            $ref: '#/definitions/server.APIStatus'
      tags:
      - Organization
  /v1/organizations/{reference}:
    delete:
      consumes:
      - application/json
      description: Delete an existing organization by ID
      parameters:
      - description: Organization ID
        in: path
        name: reference
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Organization deleted successfully
          schema:
            $ref: '#/definitions/server.APIStatus'
        "400":
          description: Invalid organization reference
          schema:
            $ref: '#/definitions/server.APIStatus'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/server.APIStatus'
        "500":
          description: Failed to delete organization
          schema:
            $ref: '#/definitions/server.APIStatus'
      tags:
      - Organization
    put:
      consumes:
      - application/json
      description: Update an existing organization by ID
      parameters:
      - description: Organization ID
        in: path
        name: reference
        required: true
        type: string
      - description: Organization update request
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/server.updateOrgRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Organization updated successfully
          schema:
            $ref: '#/definitions/server.OrganizationResponse'
        "400":
          description: Invalid organization reference or request body
          schema:
            $ref: '#/definitions/server.APIStatus'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/server.APIStatus'
        "500":
          description: Failed to update organization
          schema:
            $ref: '#/definitions/server.APIStatus'
      tags:
      - Organization
...
schemes:
- http
- https
swagger: "2.0"